DataBase & FrontEnd Update
I have added the google api for the <h1> tag. Changed some CSS and Images to get them the way that I want them. Created a Database class that is implemented in the adminlogin.php . Now all I have to do is CRUD in my admin page and Show my github projects in the projects tab!